@@ -1,6 +1,6 @@
 {
   "name": "sugar-chalk",
-  "version": "1.6.7",
+  "version": "2.0.0",
   "description": "Sugar syntax for chalk",
   "keywords": [
     "chalk",
@@ -19,42 +19,42 @@
     "url": "https://github.com/marcol/sugar-chalk.git"
   },
   "author": "Marco Dias Lopes <rhino@pacodacomenda.com>",
+  "type": "module",
   "scripts": {
     "lint": "yarn lint:js && yarn lint:md",
     "lint:js": "eslint .",
     "lint:md": "remark .",
-    "test": "jest",
+    "test": "vitest",
+    "coverage": "vitest run --coverage",
     "sandbox": "node sandbox.js"
   },
   "dependencies": {
-    "chalk": "^4.1.2"
+    "chalk": "^5.1.2"
   },
   "devDependencies": {
-    "@commitlint/cli": "^13.1.0",
-    "@commitlint/config-conventional": "^13.1.0",
+    "@commitlint/cli": "^17.2.0",
+    "@commitlint/config-conventional": "^17.2.0",
+    "@vitest/coverage-c8": "^0.25.1",
     "babel-eslint": "^10.1.0",
-    "eslint": "^7.32.0",
-    "eslint-config-standard": "^16.0.3",
+    "eslint": "^8.27.0",
+    "eslint-config-standard": "^17.0.0",
     "eslint-plugin-filenames": "^1.3.2",
-    "eslint-plugin-html": "^6.1.2",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-plugin-jest": "^24.4.2",
+    "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-json-format": "^2.0.1",
-    "eslint-plugin-markdown": "^2.2.1",
+    "eslint-plugin-markdown": "^3.0.0",
+    "eslint-plugin-n": "^15.5.1",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^5.1.0",
-    "husky": "^7.0.2",
-    "jest": "^27.2.0",
-    "remark-cli": "^10.0.0",
-    "remark-lint": "^9.0.1",
-    "remark-preset-lint-markdown-style-guide": "^5.0.1"
+    "eslint-plugin-promise": "^6.1.1",
+    "eslint-plugin-vitest": "^0.0.17",
+    "husky": "^8.0.2",
+    "remark-cli": "^11.0.0",
+    "remark-lint": "^9.1.1",
+    "remark-preset-lint-markdown-style-guide": "^5.1.2",
+    "vitest": "^0.25.1"
   },
   "resolutions": {
-    "**/trim-newlines": "^4.0.1",
-    "**/lodash": "^4.17.21",
-    "**/path-parse": "^1.0.7",
-    "**/glob-parent": "^5.1.2",
-    "**/hosted-git-info": "^3.0.8",
-    "**/ini": "^1.3.6"
+    "**/minimatch": "^3.0.5",
+    "**/ansi-regex": "^5.0.1",
+    "**/minimist": "^1.2.6"
   }
 }
