@@ -1,22 +1,23 @@
-const { info, error } = require('../index')
-const { MSGS, OBJ, ARR, ERR, INFO } = require('../__mocks__/mocks')
+import { describe, expect, it } from 'vitest'
+import { info, error } from '../index'
+import mock from '../__mocks__/mocks'
 
 process.env.TEST = true
 
 describe('tests arguments functionality', () => {
-  test('tests multiple', () => {
-    expect(info(...MSGS)).toBe(INFO + ' ' + MSGS.join(' '))
+  it('tests multiple', () => {
+    expect(info(...mock.MSGS)).toBe(mock.INFO + ' ' + mock.MSGS.join(' '))
   })
 
-  test('tests object', () => {
-    expect(info(OBJ)).toContain(JSON.stringify(OBJ))
+  it('tests object', () => {
+    expect(info(mock.OBJ)).toContain(JSON.stringify(mock.OBJ))
   })
 
-  test('test array', () => {
-    expect(info(ARR)).toContain(JSON.stringify(ARR))
+  it('test array', () => {
+    expect(info(mock.ARR)).toContain(JSON.stringify(mock.ARR))
   })
 
-  test('tests error', () => {
-    expect(error(ERR)).toContain(ERR.message)
+  it('tests error', () => {
+    expect(error(mock.ERR)).toContain(mock.ERR.message)
   })
 })
