@@ -1,10 +1,11 @@
-const { done } = require('../index')
-const { MSG, DONE } = require('../__mocks__/mocks')
+import { describe, expect, it } from 'vitest'
+import { done } from '../index'
+import mock from '../__mocks__/mocks'
 
 process.env.TEST = true
 
 describe('tests done functionality', () => {
   it('checks done', () => {
-    expect(done(MSG)).toBe(DONE + ' ' + MSG)
+    expect(done(mock.MSG)).toBe(mock.DONE + ' ' + mock.MSG)
   })
 })
