@@ -1,10 +1,11 @@
-const { warn } = require('../index')
-const { MSG, WARN } = require('../__mocks__/mocks')
+import { describe, expect, it } from 'vitest'
+import { warn } from '../index'
+import mock from '../__mocks__/mocks'
 
 process.env.TEST = true
 
 describe('tests warn functionality', () => {
   it('checks warn', () => {
-    expect(warn(MSG)).toBe(WARN + ' ' + MSG)
+    expect(warn(mock.MSG)).toBe(mock.WARN + ' ' + mock.MSG)
   })
 })
