@@ -1,9 +1,14 @@
-const { log, done, error, silent } = require('./index')
-const { MSG, ERR, OBJ } = require('./__mocks__/mocks')
+import { log, info, debug, done, pass, warn, error, fail, silent } from './index.js'
+import mock from './__mocks__/mocks.js'
 
-log(OBJ)
-error(ERR)
-done(MSG)
+log(mock.OBJ)
+info(mock.MSG)
+debug(mock.MSG)
+warn(mock.MSG)
+error(mock.MSG)
+fail(mock.MSG)
+done(mock.MSG)
+pass(mock.MSG)
 silent(true)
 log('Silent mode - should not be visible')
 silent(false)
